refactor(command-manager-lib): change command-manager root dir to remote-command-manager
